REFACTOR: Replaced Headless UI modal with Shadcn for language section in home - Modified base shadcn component significantly
